python/requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0

# PDF export
reportlab==4.0.4

# SVG export
svgwrite==1.4.3

# Development dependencies (optional)
pytest==7.4.0
black==23.7.0
flake8==6.1.0
